feat(03-01): create test_corr_patterns.py with happy-path tests for all 4 messaging patterns
- Add test_pattern_pub_sub: publish-subscribe with pre-publish and live messages
- Add test_pattern_rpc: request-reply with addition callback
- Add test_pattern_progress: progress publisher task lifecycle through completion
- Add test_pattern_journal: journal pub/read with pre-published and live messages
- Move tests from test_messenger.py, test_messenger_reqresp.py, test_messenger_journal.py
- Clean unused imports from original files
